# Change default storage layout to column store whenever the schema permits