Change the default layout to column store. This will likely slow down updates but likely speed up analytic queries. How much? Here are some numbers (Release build) on the SSB 10GB benchmark on a single socket with 10 cores. 

Query | Before  | After 
:--------|--------:|--------:|
Load the big `lineitem` table | 123.7 seconds | 207.6 seconds 
Q1 | 3.0 seconds | 0.96 seconds  

Need to work on speeding up insertion into (compressed) column stores. A future PR. 

PS: The following command was use to start quickstep: `./quickstep_cli_shell -printing_enabled=false -buffer_pool_slots=10240 -parallelize_load -num_workers=10 -worker_affinities="0,4,8,12,16,20,24,28,32,36"`