@zuyu 2016-02-11
@pateljm This is awesome!

One minor grammar issue: capitalize the first letter in all the comments, as we treat them as English sentences.