@cramja 2016-02-11
The code looks good.

There are a few corner cases to consider.

* If the user creates a table where the first attribute is an attribute which does not have a standard comparator, then setting the SORT_ID to that attribute will make an invalid description.
* Similarly, there are some attribute types which cannot be compressed in the relation, the resulting description will be invalid.

For more information on this, the [DescriptionIsValid](https://github.com/pivotalsoftware/quickstep/blob/master/storage/CompressedColumnStoreTupleStorageSubBlock.cpp#L94) method of CompressedColumnStore will show you what conditions must be met while creating a block.

* Finally, setting the default block to column store removes the ability to ad-hoc add tuples to the store. (What if we could keep a rowstore block around for ad-hoc insertions? Is the this even a useful feature?)