@cramja 2016-02-11
I should add that a quick fix to the invalid description corner case if to create the description. We can then test if it is valid, and if not, use the old method of using a rowstore.