@pateljm 2016-02-12
Excellent points @cramja on catching the case when the first attribute does not have a comparator. Let me mull over the other point you make about ad hoc tuples. 

On a different note, I'm so glad to see you ramping up so quickly and your attention to detail. Thanks!