@pateljm 2016-02-12
Ok - had a nice chat with @cramja and would make sense to not merge this PR even after reworking. Better idea is to wait for the block properties work to be completed, and then build on top of that. Deleting this PR next. Thanks Marc!